Widen buffer identity to 64 bits and refuse reuse at exhaustion
The counter's job is the claim that it never repeats, which 32 bits does
not get to make: 4.29 billion claims is a number, and a wrapped id hands
a live buffer the identity a long-dead one's handle still holds.  id and
generation become uint64_t, a claim that reaches UINT64_MAX leaves the
slot with id 0 and says so, and buf_resolve() never matches id 0 -- so a
zeroed handle, which is how a detached view will be spelled, is a
definite no.

+++ src/def.h
@@ -358,9 +358,10 @@ struct editor_buffer {
 	 * the life of the process; `generation` counts the times this slot has
 	 * changed hands.  Both are needed to tell "the buffer I meant" from
 	 * "another buffer that happens to sit where it did".  A slot that has
-	 * never been used has id 0. */
-	uint32_t id;
-	uint32_t generation;
+	 * never been used has id 0, and so does one a claim could not give an
+	 * identity to; neither is ever matched. */
+	uint64_t id;
+	uint64_t generation;
 	struct kg_point last_point; /* Where a view last left off here. */
 	int numrows;
 	erow *row;
@@ -406,8 +407,8 @@ struct editor_buffer {
  * names nothing. */
 struct kg_buffer_handle {
 	int slot;
-	uint32_t id;
-	uint32_t generation;
+	uint64_t id;
+	uint64_t generation;
 };
 
 /* Global editor state */
